Improve "Day 13: Care Package" code structure

+++ src/Tasks/2019/Day13/Tile.cs
@@ -2,8 +2,8 @@ namespace App.Tasks.Year2019.Day13
 {
     public class Tile
     {
-        public long X { get; set; }
-        public long Y { get; set; }
-        public long Id { get; set; }
+        public int X { get; set; }
+        public int Y { get; set; }
+        public int Id { get; set; }
     }
 }
